+++ packages/react/src/components/Auth/Icons.tsx
@@ -6,7 +6,33 @@ const iconDefaultStyles = css({
   height: '21px',
 })
 
-export const google = () => (
+export interface IconsProps {
+  provider: string
+}
+
+export const Icons = ({ provider }: IconsProps) => {
+  if (provider == 'google') return google()
+  if (provider == 'facebook') return facebook()
+  if (provider == 'twitter') return twitter()
+  if (provider == 'apple') return apple()
+  if (provider == 'github') return github()
+  if (provider == 'gitlab') return gitlab()
+  if (provider == 'bitbucket') return bitbucket()
+  if (provider == 'discord') return discord()
+  if (provider == 'azure') return azure()
+  if (provider == 'keycloak') return keycloak()
+  if (provider == 'linkedin') return linkedin()
+  if (provider == 'linkedin_oidc') return linkedin()
+  if (provider == 'notion') return notion()
+  if (provider == 'slack') return slack()
+  if (provider == 'spotify') return spotify()
+  if (provider == 'twitch') return twitch()
+  if (provider == 'workos') return workos()
+  if (provider == 'kakao') return kakao()
+  return null
+}
+
+const google = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -33,7 +59,7 @@ export const google = () => (
   </svg>
 )
 
-export const facebook = () => (
+const facebook = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -49,7 +75,7 @@ export const facebook = () => (
   </svg>
 )
 
-export const twitter = () => (
+const twitter = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -57,14 +83,12 @@ export const twitter = () => (
     width="21px"
     height="21px"
   >
-    <path
-      fill="#03A9F4"
-      d="M42,12.429c-1.323,0.586-2.746,0.977-4.247,1.162c1.526-0.906,2.7-2.351,3.251-4.058c-1.428,0.837-3.01,1.452-4.693,1.776C34.967,9.884,33.05,9,30.926,9c-4.08,0-7.387,3.278-7.387,7.32c0,0.572,0.067,1.129,0.193,1.67c-6.138-0.308-11.582-3.226-15.224-7.654c-0.64,1.082-1,2.349-1,3.686c0,2.541,1.301,4.778,3.285,6.096c-1.211-0.037-2.351-0.374-3.349-0.914c0,0.022,0,0.055,0,0.086c0,3.551,2.547,6.508,5.923,7.181c-0.617,0.169-1.269,0.263-1.941,0.263c-0.477,0-0.942-0.054-1.392-0.135c0.94,2.902,3.667,5.023,6.898,5.086c-2.528,1.96-5.712,3.134-9.174,3.134c-0.598,0-1.183-0.034-1.761-0.104C9.268,36.786,13.152,38,17.321,38c13.585,0,21.017-11.156,21.017-20.834c0-0.317-0.01-0.633-0.025-0.945C39.763,15.197,41.013,13.905,42,12.429"
-    />
+    <path fill="black"
+      d="M28.4714 20.3247L45.9682 0H41.822L26.6295 17.6477L14.4953 0H0.5L18.8493 26.6864L0.5 48H4.64642L20.6901 29.3635L33.5047 48H47.5L28.4704 20.3247H28.4714ZM22.7923 26.9215L20.9331 24.2642L6.14043 3.11923H12.5091L24.447 20.1839L26.3061 22.8412L41.824 45.0226H35.4553L22.7923 26.9225V26.9215Z" />
   </svg>
 )
 
-export const apple = () => (
+const apple = () => (
   <svg
     className={iconDefaultStyles()}
     fill="gray"
@@ -78,7 +102,7 @@ export const apple = () => (
   </svg>
 )
 
-export const github = () => (
+const github = () => (
   <svg
     className={iconDefaultStyles()}
     fill="gray"
@@ -92,7 +116,7 @@ export const github = () => (
   </svg>
 )
 
-export const gitlab = () => (
+const gitlab = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -109,7 +133,8 @@ export const gitlab = () => (
     <path fill="#ffa726" d="M24 43L6 20 3 28z" />
   </svg>
 )
-export const bitbucket = () => (
+
+const bitbucket = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -146,7 +171,7 @@ export const bitbucket = () => (
   </svg>
 )
 
-export const discord = () => (
+const discord = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -161,7 +186,7 @@ export const discord = () => (
   </svg>
 )
 
-export const azure = () => (
+const azure = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -227,7 +252,7 @@ export const azure = () => (
   </svg>
 )
 
-export const keycloak = () => (
+const keycloak = () => (
   <svg
     className={iconDefaultStyles()}
     width="512"
@@ -243,7 +268,7 @@ export const keycloak = () => (
   </svg>
 )
 
-export const linkedin = () => (
+const linkedin = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -262,7 +287,7 @@ export const linkedin = () => (
   </svg>
 )
 
-export const notion = () => (
+const notion = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -293,7 +318,7 @@ export const notion = () => (
   </svg>
 )
 
-export const slack = () => (
+const slack = () => (
   <svg
     className={iconDefaultStyles()}
     xmlns="http://www.w3.org/2000/svg"
@@ -336,7 +361,7 @@ export const slack = () => (
   </svg>
 )
 
-export const spotify = () => (
+const spotify = () => (
   <svg
     className={iconDefaultStyles()}
     width="512"
@@ -352,7 +377,7 @@ export const spotify = () => (
   </svg>
 )
 
-export const twitch = () => (
+const twitch = () => (
   <svg
     className={iconDefaultStyles()}
     width="512"
@@ -371,7 +396,7 @@ export const twitch = () => (
   </svg>
 )
 
-export const workos = () => (
+const workos = () => (
   <svg
     className={iconDefaultStyles()}
     width="512"
@@ -387,6 +412,26 @@ export const workos = () => (
     <path
       d="M480 256.058C480 247.539 477.684 239.021 473.516 231.883L393.849 94.6596C379.028 69.3331 347.995 56.4396 317.888 66.34C302.603 71.4053 290.329 82.6871 282.224 96.5015L264.391 127.354L339.194 256.058L260.223 392.131L241.232 424.825C235.443 434.495 228.032 442.553 219 449H227.106H340.815C362.122 449 381.807 437.718 392.46 419.299L473.284 280.003C477.684 272.866 480 264.577 480 256.058Z"
       fill="#6363F1"
+    />
+  </svg>
+)
+
+const kakao = () => (
+  <svg
+    className={iconDefaultStyles()}
+    xmlns="http://www.w3.org/2000/svg"
+    width="2500"
+    height="2500"
+    viewBox="0 0 256 256"
+  >
+    <path
+      fill="#FFE812"
+      d="M256 236c0 11.046-8.954 20-20 20H20c-11.046 0-20-8.954-20-20V20C0 8.954 8.954 0 20 0h216c11.046 0 20 8.954 20 20v216z"
+    />
+    <path d="M128 36C70.562 36 24 72.713 24 118c0 29.279 19.466 54.97 48.748 69.477-1.593 5.494-10.237 35.344-10.581 37.689 0 0-.207 1.762.934 2.434s2.483.15 2.483.15c3.272-.457 37.943-24.811 43.944-29.04 5.995.849 12.168 1.29 18.472 1.29 57.438 0 104-36.712 104-82 0-45.287-46.562-82-104-82z" />
+    <path
+      fill="#FFE812"
+      d="M70.5 146.625c-3.309 0-6-2.57-6-5.73V105.25h-9.362c-3.247 0-5.888-2.636-5.888-5.875s2.642-5.875 5.888-5.875h30.724c3.247 0 5.888 2.636 5.888 5.875s-2.642 5.875-5.888 5.875H76.5v35.645c0 3.16-2.691 5.73-6 5.73zM123.112 146.547c-2.502 0-4.416-1.016-4.993-2.65l-2.971-7.778-18.296-.001-2.973 7.783c-.575 1.631-2.488 2.646-4.99 2.646a9.155 9.155 0 0 1-3.814-.828c-1.654-.763-3.244-2.861-1.422-8.52l14.352-37.776c1.011-2.873 4.082-5.833 7.99-5.922 3.919.088 6.99 3.049 8.003 5.928l14.346 37.759c1.826 5.672.236 7.771-1.418 8.532a9.176 9.176 0 0 1-3.814.827c-.001 0 0 0 0 0zm-11.119-21.056L106 108.466l-5.993 17.025h11.986zM138 145.75c-3.171 0-5.75-2.468-5.75-5.5V99.5c0-3.309 2.748-6 6.125-6s6.125 2.691 6.125 6v35.25h12.75c3.171 0 5.75 2.468 5.75 5.5s-2.579 5.5-5.75 5.5H138zM171.334 146.547c-3.309 0-6-2.691-6-6V99.5c0-3.309 2.691-6 6-6s6 2.691 6 6v12.896l16.74-16.74c.861-.861 2.044-1.335 3.328-1.335 1.498 0 3.002.646 4.129 1.772 1.051 1.05 1.678 2.401 1.764 3.804.087 1.415-.384 2.712-1.324 3.653l-13.673 13.671 14.769 19.566a5.951 5.951 0 0 1 1.152 4.445 5.956 5.956 0 0 1-2.328 3.957 5.94 5.94 0 0 1-3.609 1.211 5.953 5.953 0 0 1-4.793-2.385l-14.071-18.644-2.082 2.082v13.091a6.01 6.01 0 0 1-6.002 6.003z"
     />
   </svg>
 )