+++ packages/svelte/src/lib/Auth/Auth.svelte
@@ -34,7 +34,8 @@
 	export let theme: 'default' | string = 'default';
 	export let localization: { variables?: I18nVariables } = {};
 	export let otpType: OtpType = 'email';
-	export let additionalData: { [key: string]: any } | undefined;
+	export let passwordLimit: boolean = false;
+	export let additionalData: { [key: string]: any } | undefined = undefined;
 
 	onMount(() => {
 		const { data: authListener } = supabaseClient.auth.onAuthStateChange((event) => {
@@ -108,18 +109,33 @@
 				{magicLink}
 				{showLinks}
 				{additionalData}
+				{passwordLimit}
 				{i18n}><slot /></EmailAuth
 			>
 		{/if}
 	{/if}
 	{#if view === VIEWS.FORGOTTEN_PASSWORD}
-		<ForgottenPassword {i18n} {supabaseClient} bind:authView={view} {showLinks} {appearance} />
+		<ForgottenPassword
+			{i18n}
+			{supabaseClient}
+			bind:authView={view}
+			{showLinks}
+			{appearance}
+			{redirectTo}
+		/>
 	{/if}
 	{#if view === VIEWS.MAGIC_LINK}
 		<MagicLink {i18n} {supabaseClient} bind:authView={view} {appearance} {redirectTo} {showLinks} />
 	{/if}
 	{#if view === VIEWS.UPDATE_PASSWORD}
-		<UpdatePassword {i18n} {supabaseClient} bind:authView={view} {appearance} {showLinks} />
+		<UpdatePassword
+			{i18n}
+			{supabaseClient}
+			bind:authView={view}
+			{appearance}
+			{passwordLimit}
+			{showLinks}
+		/>
 	{/if}
 	{#if view === VIEWS.VERIFY_OTP}
 		<VerifyOtp {i18n} {supabaseClient} bind:authView={view} {appearance} {showLinks} {otpType} />