+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @supabase/auth-ui-shared
 
+## 0.1.8
+
+### Patch Changes
+
+- a8f9395d: Message styling change
+
+## 0.1.7
+
+### Patch Changes
+
+- 2b252910: Fix check for blank email address on magic links
+
 ## 0.1.6
 
 ### Patch Changes